+++ src/search.cpp
@@ -70,7 +70,7 @@ void Searcher::wait() {
     std::unique_lock lock_guard{mutex};
 }
 
-void Searcher::initialize(int thread_count) {
+void Searcher::initialize(size_t thread_count) {
     if (m_workers.size() == thread_count) {
         return;
     }
@@ -88,7 +88,7 @@ void Searcher::initialize(int thread_count) {
 
     if (thread_count > 0) {
         m_workers.push_back(std::make_unique<Worker>(*this, ThreadType::MAIN));
-        for (int i = 1; i < thread_count; i++) {
+        for (size_t i = 1; i < thread_count; i++) {
             m_workers.push_back(std::make_unique<Worker>(*this, ThreadType::SECONDARY));
         }
     }
@@ -342,7 +342,7 @@ Value Worker::search(
     // Draw checks
     if (!ROOT_NODE) {
         // Repetition check
-        if (repetition_info.detect_repetition(ply)) {
+        if (repetition_info.detect_repetition(static_cast<usize>(ply))) {
             return 0;
         }
         // 50 mr check
@@ -399,7 +399,8 @@ Value Worker::search(
 
     if (!PV_NODE && !is_in_check && !pos.is_kp_endgame() && depth >= tuned::nmp_depth
         && tt_adjusted_eval >= beta) {
-        int      R = tuned::nmp_base_r + depth / 4 + std::min(3, (tt_adjusted_eval - beta) / 400) + improving;
+        int R =
+          tuned::nmp_base_r + depth / 4 + std::min(3, (tt_adjusted_eval - beta) / 400) + improving;
         Position pos_after = pos.null_move();
 
         repetition_info.push(pos_after.get_hash_key(), true);
